+++ app/models/user.rb
@@ -5,4 +5,6 @@ class User < ApplicationRecord
   # :rememberable, :lockable, :timeoutable, :trackable and :omniauthable
   devise :database_authenticatable, :registerable, :confirmable,
          :recoverable, :validatable
+
+  has_many :accounts, class_name: 'Users::Account', inverse_of: :user, dependent: :destroy
 end