+++ db/migrate/20201115162247_create_users_accounts.rb
@@ -1,0 +1,10 @@
+class CreateUsersAccounts < ActiveRecord::Migration[6.0]
+  def change
+    create_table :users_accounts do |t|
+      t.integer :user_id
+      t.string :name, limit: 255
+      t.timestamps
+    end
+    add_index :users_accounts, :user_id
+  end
+end