+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bond model
 - Foundation model
 - Exchanges::Quote model
+- Users::Account model